TIP-841: Move the integration test to Akeneo\Channel
